# app/server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.BookshelfGroupItem:
    properties:
      bookCount:
        type: integer
      groupName:
        type: string
    type: object
  dto.BookshelfPageResponse:
    properties:
      current:
        type: integer
      records:
        items:
          $ref: '#/definitions/dto.BookshelfResponse'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  dto.BookshelfRequest:
    properties:
      bookId:
        type: integer
      groupName:
        description: 可选，默认"默认"
        type: string
    required:
    - bookId
    type: object
  dto.BookshelfResponse:
    properties:
      addTime:
        type: string
      bookAuthor:
        type: string
      bookCover:
        type: string
      bookId:
        type: integer
      bookTitle:
        type: string
      chapterId:
        type: integer
      chapterNo:
        type: integer
      createBy:
        type: integer
      createTime:
        type: string
      groupName:
        type: string
      id:
        type: integer
      isTop:
        type: boolean
      lastReadTime:
        type: string
      position:
        type: integer
      readDuration:
        type: integer
      readPercent:
        type: number
      readerId:
        type: integer
      totalChapters:
        type: integer
      totalWords:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.BookshelfSearch:
    properties:
      current:
        type: integer
      groupName:
        description: 按分组筛选
        type: string
      keyword:
        description: 按书名关键词搜索
        type: string
      size:
        type: integer
    type: object
  dto.BookshelfUpdateRequest:
    properties:
      groupName:
        description: 修改分组
        type: string
      isTop:
        description: 修改置顶状态
        type: boolean
    type: object
...
  dto.ReadEventRequest:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      deviceType:
        description: 默认 web
        type: string
      durationSec:
        description: 本次心跳区间阅读时长(秒)
        type: integer
      sessionId:
        description: 服务端生成 UUID, 客户端可传空
        type: string
      wordCount:
        description: 本次心跳区间滚动字数
        type: integer
    required:
    - bookId
    - chapterId
    - durationSec
    type: object
...
  dto.ReadProgressRequest:
    properties:
      chapterId:
        type: integer
      chapterNo:
        type: integer
      fileId:
        description: 文件切换时传入
        type: integer
      percent:
        type: number
      position:
        type: integer
      readDuration:
        description: 本次增加的阅读时长(秒)
        type: integer
    required:
    - chapterId
    - chapterNo
    type: object
  dto.ReadProgressResponse:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      chapterNo:
        type: integer
      createBy:
        type: integer
      createTime:
        type: string
      fileId:
        type: integer
      id:
        type: integer
      lastReadTime:
        type: string
      percent:
        type: number
      position:
        type: integer
      readDuration:
        type: integer
      readerId:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 小说阅读平台后端 API
  title: Boread API
  version: "1.0"
paths:
...
  /api/book/reader/progress/{bookId}:
    get:
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReadProgressResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取阅读进度
      tags:
      - book-reader
    put:
      consumes:
      - application/json
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      - description: 进度参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReadProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReadProgressResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 上报阅读进度
      tags:
      - book-reader
  /api/book/reader/read-event:
    post:
      consumes:
      - application/json
      parameters:
      - description: 阅读事件参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReadEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 上报阅读事件
      tags:
      - book-reader
...
  /api/book/shelf:
    post:
      consumes:
      - application/json
      parameters:
      - description: 书架参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BookshelfRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookshelfResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 添加到书架
      tags:
      - reader-bookshelf
  /api/book/shelf/{bookId}:
    delete:
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 从书架移除
      tags:
      - reader-bookshelf
    put:
      consumes:
      - application/json
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BookshelfUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookshelfResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新书架
      tags:
      - reader-bookshelf
  /api/book/shelf/groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BookshelfGroupItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取书架分组列表
      tags:
      - reader-bookshelf
  /api/book/shelf/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BookshelfSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookshelfPageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 书架分页列表
      tags:
      - reader-bookshelf
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"